feat(screens): Use walletMetadataLoader and shimmer network icon to show wallet icon
