+++ lib/services/chain_loader.dart
@@ -1,3 +1,6 @@
+// Copyright (c) Silence Laboratories Pte. Ltd.
+// This software is licensed under the Silence Laboratories License Agreement.
+
 import 'dart:convert';
 import 'package:flutter/services.dart';
 import 'package:silentshard/screens/sign/sign_request_view_model.dart';
